backend/requirements.txt
Django>=4.2,<5.0
djangorestframework>=3.14
djangorestframework-simplejwt>=5.2
django-cors-headers>=4.0